src/main/java/com/totvs/tj/tcc/domain/conta/Conta.java
package com.totvs.tj.tcc.domain.conta;

import static com.totvs.tj.tcc.domain.conta.Conta.Situacao.ABERTO;
import static com.totvs.tj.tcc.domain.conta.Conta.Situacao.SUSPENSO;
import static lombok.AccessLevel.PRIVATE;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.ToString;

@Getter
@ToString
@Builder
@AllArgsConstructor(access = PRIVATE)
public class Conta {

    private ContaId id;
    
    private Empresa empresa;
    
    private ResponsavelId responsavel;

    private Situacao situacao;
    
    public void suspender() {
        situacao = SUSPENSO;
    }

    public boolean isDisponivel() {
        return ABERTO.equals(situacao);
    }    
   
    static enum Situacao {

        ABERTO, SUSPENSO;
        
    }
}
